Add presence to slack connection
