+++ src/SlackLibrary/Connections/Sockets/Messages/Outbound/PresenceSubMessage.cs
@@ -1,0 +1,18 @@
+﻿using Newtonsoft.Json;
+using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace SlackLibrary.Connections.Sockets.Messages.Outbound
+{
+	public class PresenceSubMessage : BaseMessage
+	{
+		[JsonProperty("ids")]
+		public IEnumerable<string> Ids { get; set; }
+
+		public PresenceSubMessage()
+		{
+			Type = "presence_sub";
+		}
+	}
+}
