+++ src/SlackLibrary/EventHandlers/PresenceChangeHandler.cs
@@ -1,0 +1,9 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace SlackLibrary.EventHandlers
+{
+	public delegate Task PresenceChangeHandler(string slackId, string presence);
+}
